TGL-34.fix: added types folder, added npm run format for prettier in package.json
