+++ src/types/typeOrder.d.ts
@@ -1,0 +1,6 @@
+export interface Order {
+  _id: string;
+  trackerNumber: number;
+  PUP: string;
+  price: number;
+}
