+++ src/types/typeUser.d.ts
@@ -1,0 +1,14 @@
+export interface User {
+  _id: string;
+  email: string;
+  marketId: string;
+  firstName: string;
+  lastName: string;
+  middleName: string;
+  role: string;
+  token: string;
+  phoneNumber: string;
+  region: string;
+  settlement: string;
+  address: string;
+}
